feat(clipmenu): detect rofi script mode
by probing for the `ROFI_RETV` environment variable.

BREAKING CHANGE: previously, if `CM_LAUNCHER` was (a) unset or (b) set
to a value other than `dmenu`, `fzf`, `rofi`, or `rofi-script`, then
clipmenu would attempt to execute the launcher, treating it as
dmenu-compatible.  Now, if conditions (a) or (b) hold, AND the
`ROFI_RETV` variable is set, clipmenu assumes it is being run in rofi
script mode and will not attempt to execute a launcher.