clipmenud: Prevent infinite loop in signal handler
When handling an enable/disable signal, clipmenud could enter an
infinite loop if another X event was already in the queue. The function
to get the server time would repeatedly read and put back the same event
at the head of the queue (er, whoops!).

Instead use XIfEvent, which doesn't bodge the event queue.

Fixes #255.